Add base name to bundles map
We need to pass the base name for the cached bundles map, so we can
identify the entries in the cache. ResourceBundle don't have a getter
for the base name.

We create a cache key to have a key that identifies the different
binaries cache elements in a shared cache.